Jenn/fix navbar active menu mismatch (#35)
* Abstracted Header into components folder
* Fixed problem where active-menu button would not match route as a result of manually entering the route instead of pressing the buttons
