+++ frontend/src/App.less
@@ -5,25 +5,6 @@ body {
   margin: 0;
 }
 
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-left: auto;
-  background-color: @purple-5;
-  height: 70px;
-}
-
-.header-content {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-}
-.logo {
-  display: flex;
-  flex-direction: row;
-}
-
 .dark {
   background-color: #1d1f20;
   color: white;
