+++ frontend/src/components/header/header.less
@@ -1,0 +1,23 @@
+@import "~antd/lib/style/themes/default.less";
+
+.header {
+    display: flex !important;
+    flex-direction: row !important; 
+    justify-content: space-between;
+    align-items: center;
+    margin-left: auto;
+    padding-left: 20px;
+    padding-right: 20px;
+    background-color: @purple-5 !important;
+    height: 70px;
+  }
+
+.header-content {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+} 
+.logo {
+  display: flex;
+  flex-direction: row;
+}
